requirements.txt
psutil~=6.1.1
python-vlc==3.0.21203

PySide6~=6.8.1.1
pyqtgraph~=0.13.7
requests~=2.32.3